% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpu.R
\name{cgv_gpu_available}
\alias{cgv_gpu_available}
\title{Check if Vulkan GPU is Available}
\usage{
cgv_gpu_available()
}
\value{
Logical
}
\description{
Checks whether ggmlR is installed and a Vulkan device is present.
}
\examples{
cgv_gpu_available()
}
