% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{cgv_run}
\alias{cgv_run}
\title{Run the Viewer Event Loop}
\usage{
cgv_run(viewer)
}
\arguments{
\item{viewer}{External pointer returned by \code{cgv_viewer}.}
}
\description{
Starts the rendering loop. Blocks until the window is closed.
}
